python-core/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
ffmpeg-python==0.2.0
pydantic==2.5.2
requests==2.31.0
urllib3<2.0.0